Migrate anonymous class test to addSourceLines
